+++ src/main/java/es/in2/walletdata/facade/impl/UserDataFacadeServiceImpl.java
@@ -5,10 +5,10 @@ import es.in2.walletdata.domain.UserEntity;
 import es.in2.walletdata.domain.UserRequest;
 import es.in2.walletdata.domain.VcBasicData;
 import es.in2.walletdata.facade.UserDataFacadeService;
-import es.in2.walletdata.service.BrokerAdapterService;
+import es.in2.walletdata.service.BrokerService;
 import es.in2.walletdata.service.UserDataService;
 import es.in2.walletdata.service.WalletCryptoService;
-import es.in2.walletdata.utils.DidMethods;
+import es.in2.walletdata.domain.DidMethods;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -23,17 +23,17 @@ import java.util.List;
 public class UserDataFacadeServiceImpl implements UserDataFacadeService {
 
     private final UserDataService userDataService;
-    private final BrokerAdapterService brokerAdapterService;
+    private final BrokerService brokerService;
     private final WalletCryptoService walletCryptoService;
 
     @Override
     public Mono<Void> saveVerifiableCredentialByUserId(String userId, String vcJwt) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId).flatMap(userEntity ->
+        return brokerService.getUserEntityFromContextBroker(userId).flatMap(userEntity ->
                 // Save the Verifiable Credential to the UserEntity
                 userDataService.saveVC(userEntity, vcJwt)).flatMap(updatedUserEntity ->
                 // Update the UserEntity back in the Context Broker
-                brokerAdapterService.updateUserEntityInContextBroker(updatedUserEntity, userId)).doOnSuccess(aVoid -> log.info("Verifiable Credential saved and UserEntity updated successfully for userId: {}", userId)).onErrorResume(e -> {
+                brokerService.updateUserEntityInContextBroker(updatedUserEntity, userId)).doOnSuccess(aVoid -> log.info("Verifiable Credential saved and UserEntity updated successfully for userId: {}", userId)).onErrorResume(e -> {
             log.error("Error in saving Verifiable Credential for userId: " + userId, e);
             return Mono.error(e);
         });
@@ -42,7 +42,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<List<VcBasicData>> getUserVCs(String userId) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId)
+        return brokerService.getUserEntityFromContextBroker(userId)
                 // Once the UserEntity is retrieved, use the userDataService to get VCs in JSON format
                 .flatMap(userDataService::getUserVCsInJson).doOnSuccess(vcBasicDataDTOList -> log.info("Retrieved VCs in JSON for userId: {}", userId)).onErrorResume(e -> {
                     log.error("Error in retrieving VCs in JSON for userId: {}", userId, e);
@@ -53,7 +53,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<Void> deleteVerifiableCredentialById(String credentialId, String userId) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId).flatMap(userEntity ->
+        return brokerService.getUserEntityFromContextBroker(userId).flatMap(userEntity ->
                 // Extract DID from the Verifiable Credential
                 userDataService.extractDidFromVerifiableCredential(userEntity, credentialId).flatMap(did ->
                         // Delete the Private Key associated with the DID
@@ -64,7 +64,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
             return userDataService.deleteVerifiableCredential(userEntity, credentialId, did);
         }).flatMap(updatedUserEntity ->
                 // Update the UserEntity back in the Context Broker
-                brokerAdapterService.updateUserEntityInContextBroker(updatedUserEntity, userId)).doOnSuccess(aVoid -> log.info("Verifiable Credential deleted and UserEntity updated successfully for userId: {}", userId)).onErrorResume(e -> {
+                brokerService.updateUserEntityInContextBroker(updatedUserEntity, userId)).doOnSuccess(aVoid -> log.info("Verifiable Credential deleted and UserEntity updated successfully for userId: {}", userId)).onErrorResume(e -> {
             log.error("Error in deleting Verifiable Credential for userId: " + userId, e);
             return Mono.error(e);
         });
@@ -73,7 +73,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<List<VcBasicData>> getVCsByVcTypeList(String userId, List<String> vcTypeList) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId)
+        return brokerService.getUserEntityFromContextBroker(userId)
                 // Once the UserEntity is retrieved, pass it to the userDataService
                 .flatMap(userEntity -> userDataService.getSelectableVCsByVcTypeList(vcTypeList, userEntity)).doOnSuccess(vcBasicDataDTOList -> log.info("Selectable VCs retrieved for userId: {}", userId)).onErrorResume(e -> {
                     log.error("Error retrieving selectable VCs for userId: {}", userId, e);
@@ -84,11 +84,11 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<Void> saveDidByUserId(String userId, String did, DidMethods didMethod) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId).flatMap(userEntity ->
+        return brokerService.getUserEntityFromContextBroker(userId).flatMap(userEntity ->
                 // Save the Did to the UserEntity
                 userDataService.saveDid(userEntity, did, didMethod)).flatMap(updatedUserEntity ->
                 // Update the UserEntity back in the Context Broker
-                brokerAdapterService.updateUserEntityInContextBroker(updatedUserEntity, userId)).doOnSuccess(aVoid -> log.info("Did saved and UserEntity updated successfully for userId: {}", userId)).onErrorResume(e -> {
+                brokerService.updateUserEntityInContextBroker(updatedUserEntity, userId)).doOnSuccess(aVoid -> log.info("Did saved and UserEntity updated successfully for userId: {}", userId)).onErrorResume(e -> {
             log.error("Error in saving Did for userId: " + userId, e);
             return Mono.error(e);
         });
@@ -97,7 +97,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<List<String>> getDidsByUserId(String userId) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId)
+        return brokerService.getUserEntityFromContextBroker(userId)
                 // Use flatMap to transform the Mono<UserEntity> to a Mono<List<String>>
                 .flatMap(userDataService::getDidsByUserEntity).doOnSuccess(dids -> log.info("Retrieved DIDs for userId: {}", userId)).onErrorResume(e -> {
                     log.error("Error in retrieving DIDs for userId: {}", userId, e);
@@ -108,7 +108,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<Void> createUserEntity(UserRequest userRequest) {
         // Create the UserEntity using the provided DTO
-        return userDataService.createUserEntity(userRequest).flatMap(brokerAdapterService::storeUserInContextBroker).doOnSuccess(aVoid -> log.info("UserEntity successfully persisted for: {}", userRequest.userId())).onErrorResume(e -> {
+        return userDataService.createUserEntity(userRequest).flatMap(brokerService::storeUserInContextBroker).doOnSuccess(aVoid -> log.info("UserEntity successfully persisted for: {}", userRequest.userId())).onErrorResume(e -> {
             log.error("Error while persisting UserEntity for: {}", userRequest.userId(), e);
             return Mono.error(e); // Propagate the error
         });
@@ -117,7 +117,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<UserAttribute> getUserDataByUserId(String userId) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId)
+        return brokerService.getUserEntityFromContextBroker(userId)
                 // Once the UserEntity is retrieved, use the userDataService to get the UserData
                 .flatMap(userDataService::getUserDataFromUserEntity).doOnSuccess(userData -> log.debug("Fetched user data for userId: {}", userId)).onErrorResume(e -> {
                     log.error("Error retrieving user data for userId: {}", userId, e);
@@ -128,7 +128,7 @@ public class UserDataFacadeServiceImpl implements UserDataFacadeService {
     @Override
     public Mono<String> getVerifiableCredentialByIdAndFormat(String userId, String credentialId, String format) {
         // Retrieve the UserEntity from the Context Broker
-        return brokerAdapterService.getUserEntityFromContextBroker(userId)
+        return brokerService.getUserEntityFromContextBroker(userId)
                 // Once the UserEntity is retrieved, use the userDataService to get the UserData
                 .flatMap(userEntity -> userDataService.getVerifiableCredentialByIdAndFormat(userEntity, credentialId, format)).doOnSuccess(userData -> log.debug("Fetched user data for userId: {}", userId)).onErrorResume(e -> {
                     log.error("Error retrieving user data for userId: {}", userId, e);